TODO: Not use axios in GET(index) request(part 1)
The data is obtained directly from the view, not from axios. The sections where this were better are Permissions, Tenders and Advertisements. Other sections is better with the pagination, just statistics sections are missing.

TODO: Get the data from the index from the view(all Controllers and Views), not from Axios(the HTTP client).